backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
chromadb==0.5.3
Pillow==10.3.0
torch==2.3.0
transformers==4.41.1
sentence-transformers==3.0.0
pydantic==2.7.1
python-dotenv==1.0.1